Split selections into plot_type and ensembles
