+++ webviz_subsurface/plugins/_rft_plotter/_views/_sim_vs_obs_view/_settings/__init__.py
@@ -1,2 +1,3 @@
-from ._selections import PlotType, Selections
+from ._plot_type import PlotType, PlotTypeSettings
+from ._selections import Selections
 from ._size_color_settings import SizeColorSettings
